Separate WasmInstance and WasmModule.

+++ dataset/GitHub_Java/oracle.graal/wasm/src/org.graalvm.wasm/src/org/graalvm/wasm/predefined/testutil/CompareContextsNode.java
@@ -50,7 +50,7 @@ import org.graalvm.wasm.WasmVoidResult;
 import org.graalvm.wasm.exception.WasmExecutionException;
 import org.graalvm.wasm.memory.WasmMemory;
 import org.graalvm.wasm.predefined.WasmBuiltinRootNode;
-import org.graalvm.wasm.predefined.testutil.SaveContextNodeNode.ContextState;
+import org.graalvm.wasm.predefined.testutil.SaveContextNode.ContextState;
 
 /**
  * Records the context state (memory and global variables) into a custom object.
@@ -112,7 +112,7 @@ public class CompareContextsNode extends WasmBuiltinRootNode {
             byte first = (byte) firstMemory.load_i32_8s(this, ptr);
             byte last = (byte) lastMemory.load_i32_8s(this, ptr);
             if (first != last) {
-                int from = (ptr - 100) / 8 * 8;
+                long from = (ptr - 100) / 8 * 8;
                 throw new WasmExecutionException(this, "Memory mismatch.\n" +
                                 "-- Reference --\n" + firstMemory.hexView(from, 200) + "\n" +
                                 "-- Actual --\n" + firstMemory.hexView(from, 200) + "\n");
